requirements.txt
openai-whisper
watchdog
textblob
pyyaml
